Use pixelsize instead of size
Font rendering is rather peculiar when the size of the font is
~10 pixels! I was surprised to learn that fonts of roughly the
same phisical size (as specified by "size=" xft parameter) seem
smaller on lower-resonution screens (in my case the ratio of
resolutions is 134/91=1.5). So it turns out that the pixesize,
which depends on resolution, is actually more suiteble than the
"physical", resolution-independend but rendering-dependent size!

Again, this is true for relatively small font sizes with
pronounced pixelation :).